refactor(UserChallengesWidget): creating selector for widget data
